crawlPage requires HTTP code 200
We consider any response whose status isn't 200 to be an error. While
all the codes that start with "2" are technically successful, it's not
yet clear to me whether we should consider all of them to indicate
success in the context of this project.

I may revisit the question of which status codes to ignore later on when
it's clear how this function will be used in the context of the project.

The boot.dev exercise suggests that `crawlPage()` should print errors as
it encounters them.

Interestingly, the authors of the boot.dev exercise has abandoned test
driven development by this point. Testing printing introduces some
complexity, requiring `console.log()` or `process.stdout.write()` to be
mocked.

Instead, I've decided to be guided by this friction, injecting a
function whose responsibility it is to decide what to do with each
error message. Error handling can then be configured at runtime, and
there's no need to mock the output streams.

I feel this is better design.