Pass site's "base" URL to crawlPage function
In order to determine whether to retrieve a link in a document we're
going to need to check whether or not it's a link to the site that we're
analysing. So we need to know two URLs; the site's base, and the link
that we're considering following.

We'll also need to be able to convert a relative link (i.e. just a path,
with no hostname) to an abolute one.