Convert JavaScript to TypeScript
The most interesting impact of this change is in the tests. We're using
mocks to control the behaviour of `fetch()`, and I'd stubbed out
response objects with object literals that were a similar shape to real
responses.

TypeScript wasn't happy, and the solution that felt the most sensible
was to use real `Response` objects instead. I thought this was
interesting; had I been more familiar with working with the `fetch()`
API would I have gone that way in the first place? Probably.